+++ src/routes.ts
@@ -3,5 +3,5 @@ import { route } from "@react-router/dev/routes"
 
 export default [
   route("/", "./routes/meetings-filtered.tsx"),
-  route("/meetings/:slug", "./routes/meeting-details.tsx"),
+  route("/group-info/:slug", "./routes/group-info.tsx"),
 ] satisfies RouteConfig
